pyproject.toml
[project]
name = "amy-lm"
version = "0.1.0"
description = "Amy LM"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cython>=3.2.2",
    "datasets>=2.15.0",
    "ipykernel>=7.1.0",
    "lightning>=2.6.0",
    "packaging>=24.2",
    "setuptools>=80.9.0",
    "torchaudio>=2.9.1",
    "torchcodec>=0.9.1",
    "wheel>=0.45.1",
]

[project.optional-dependencies]
# Machine A: Alignment (System deps: ffmpeg, libsndfile1)
alignment = [
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "soundfile>=0.12.0",
    "accelerate>=0.20.0",
    "nemo_toolkit[asr]"
]
wavlm = [
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "soundfile>=0.12.0",
    "numpy>=1.24.0",
]
llm = [
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "accelerate",
    "sentencepiece",
    "protobuf",
]
upload = [
    "huggingface_hub>=0.20.0",
]
